+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 10.8.4
 - Fix save multiple SelectInput() when none selected, add hidden empty input in Inputs.php
 - Fix check if Available Seats < selected students calculation in MassSchedule.php
 - SQL order Requests by Course in Requests.php
+- CSS set minimum colorBox width to 50% view width in StudentGrades.php
 
 Changes in 10.8.3
 -----------------